metodo comprar hecho pero hay un fallo en el metodo de añadir vehiculos

+++ src/database.h
@@ -8,7 +8,7 @@ int crearTablaUsuarios(sqlite3* db);
 int registrarUsuario(sqlite3* db, const char* nombre_usuario, const char* contrasena, const char* email);
 int verificarUsuario(sqlite3* db, const char* nombre_usuario, const char* contrasena);
 int crearTablaVentas(sqlite3 *db);
-int registrarVenta(sqlite3 *db,int id_usuario, int id_vehiculo, double precio_final);
+int registrarVenta(sqlite3 *db,int id_usuario, int id_vehiculo, int precio_final);
 int crearTablaVehiculos(sqlite3 *db);
 int registrarVehiculo(sqlite3 *db,char *marca, char *modelo, int anio, int precio);
 
@@ -17,6 +17,6 @@ int vehiculoExiste(sqlite3 *db, const char *marca, const char *modelo, int anio)
 int crearTablaPlantilla(sqlite3 *db);
 int mostrarPlantilla(sqlite3 *db);
 float obtenerPrecioVehiculo(sqlite3 *db, const char *marca, const char *modelo);
-int buscarIDVehiculo(sqlite3 *db,const char *marca, const char *modelo, int anio,float precio);
+int buscarIDVehiculo(sqlite3 *db,const char *marca, const char *modelo, int anio,int precio);
 int buscarIDUsuario(sqlite3 *db,const char *nombre_usuario );
 #endif
